foodgram/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.4.0
weasyprint==57.2
